Allow for integer application IDs:
Since GitHub application IDs are always integers, it would be nice if the
`AppInstallationAuthStrategy` (and similar) classes allowed for `int` types in
addition to `str`. We could just replace `str` with `int`, but that wouldn't be
backwards compatible, so I decided to just keep both. I can update that though
if using `int` only is preferred.